remove clean up code for transition, not necessary because the state change in the trie is not committed

+++ pkg/dex/account.go
@@ -7,8 +7,8 @@ const (
 )
 
 type Balance struct {
-	Available int
-	Pending   int
+	Available uint64
+	Pending   uint64
 }
 
 type Account struct {
